move CallableBool to django.utils.deprecation
