+++ django/contrib/auth/models.py
@@ -10,8 +10,8 @@ from django.core.mail import send_mail
 from django.db import models
 from django.db.models.manager import EmptyManager
 from django.utils import six, timezone
+from django.utils.deprecation import CallableFalse, CallableTrue
 from django.utils.encoding import python_2_unicode_compatible
-from django.utils.functional import CallableFalse, CallableTrue
 from django.utils.translation import ugettext_lazy as _
